+++ test/sql/adm1.sql
@@ -79,9 +79,9 @@ set role emaj_regression_tests_adm_user;
 -----------------------------
 -- recreate and start groups
 -----------------------------
--- set the parameter to drop the emaj_user_port column and add an emaj_appname column
+-- set the parameter to drop the emaj_user_port column and add an 'extra_col_appname' column
 insert into emaj.emaj_param (param_key, param_value_text) values ('alter_log_table',
-  'DROP COLUMN emaj_user_port, ADD COLUMN emaj_appname TEXT DEFAULT current_setting(''application_name'')');
+  'DROP COLUMN emaj_user_port, ADD COLUMN extra_col_appname TEXT DEFAULT current_setting(''application_name'')');
 
 select emaj.emaj_create_group('myGroup1');
 select emaj.emaj_comment_group('myGroup1','This is group #1');
@@ -147,15 +147,15 @@ select * from mySchema1.myTbl2b order by col20;
 select col31,col33 from mySchema1."myTbl3" order by col31;
 select * from mySchema1.myTbl4 order by col41;
 -- log tables
-select col11, col12, col13, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, emaj_appname
+select col11, col12, col13, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, extra_col_appname
     from emaj.mySchema1_myTbl1_log order by emaj_gid, emaj_tuple desc;
-select col21, col22, col23, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, emaj_appname
+select col21, col22, col23, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, extra_col_appname
     from emaj.mySchema1_myTbl2_log order by emaj_gid, emaj_tuple desc;
-select col20, col21, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, emaj_appname
+select col20, col21, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, extra_col_appname
     from emajb.mySchema1_myTbl2b_log order by emaj_gid, emaj_tuple desc;
-select col31, col33, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, emaj_appname
+select col31, col33, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, extra_col_appname
     from "emajC"."myschema1_myTbl3_log" order by emaj_gid, emaj_tuple desc;
-select col41, col42, col43, col44, col45, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, emaj_appname
+select col41, col42, col43, col44, col45, emaj_verb, emaj_tuple, emaj_gid, emaj_user, emaj_user_ip, extra_col_appname
     from emaj.mySchema1_myTbl4_log order by emaj_gid, emaj_tuple desc;
 -----------------------------
 -- Step 2 : for myGroup2, start, update tables and set 2 marks